Generalize libuv helper functions
