+++ binding.gyp
@@ -18,6 +18,7 @@
             "src/polling/polled_root.cpp",
             "src/polling/polling_iterator.cpp",
             "src/polling/polling_thread.cpp",
+            "src/helper/libuv.cpp",
             "src/nan/all_callback.cpp",
             "src/nan/functional_callback.cpp",
             "src/nan/options.cpp"
